refactor: NextButton 공용 컴포넌트로 변경
